refactor: split section
